Move fetching to _layout
For some reason, preload() of the index route would be called every time we change any component's state. Instead, we move the preload() behavior to _layout to (hopefully) make sure it loads only once.
