+++ src/routes/index.svelte
@@ -1,45 +1,6 @@
-<script context="module">
-  export async function preload(page, session) {
-    const userId = "1";
-    const userData = await this.fetch(`cards/${userId}.json`)
-      .then((res) => {
-        if (res.ok) {
-          return res.json();
-        } else {
-          this.error(res.status, res.json().message);
-        }
-      })
-      .catch((err) => {
-        this.error(err);
-      });
-
-    console.debug(`[index.svelte] userData: ${JSON.stringify(userData)}`);
-
-    console.debug("[index.svelte] Attemping POST...");
-    console.debug(`[index.svelte] POST body: ${JSON.stringify(userData)}`);
-
-    await this.fetch(`cards/${userId}.json`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(userData),
-    }).then((post_res) => {
-      if (post_res.ok) {
-        console.debug(`[index.svelte] POST success!`);
-        session.user_id = userData.user_id;
-        session.lists = userData.lists;
-      } else {
-        console.debug(`[index.svelte] POST failed.`);
-        this.error(post_res.status, post_res.message);
-        // throw new Error(`POST response failed: ${JSON.stringify(r)}`);
-      }
-    });
-  }
-</script>
-
 <script>
   import { stores } from "@sapper/app";
+  import { Button } from "sveltestrap/src";
   import Board from "../components/Board.svelte";
 
   const { session } = stores();
